[S07P31A701-136][BE] feat: boardLike, commentLike API

+++ BackEnd/youngclimb/src/main/java/com/youngclimb/domain/model/dto/board/BoardCreate.java
@@ -11,10 +11,10 @@ import java.time.LocalDateTime;
 @Builder
 public class BoardCreate {
     Long memberId;
-    Long centerId;
-    Long centerLevelId;
+    Integer centerId;
+    Integer centerLevelId;
     Long mediaId;
-    Long wallId;
+    Integer wallId;
     String level;
     String holdColor;
     LocalDate solvedDate;
