+++ BackEnd/youngclimb/src/main/java/com/youngclimb/domain/model/repository/CenterLevelRepository.java
@@ -3,5 +3,5 @@ package com.youngclimb.domain.model.repository;
 import com.youngclimb.domain.model.entity.CenterLevel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CenterLevelRepository extends JpaRepository<CenterLevel, Long> {
+public interface CenterLevelRepository extends JpaRepository<CenterLevel, Integer> {
 }
