+++ BackEnd/youngclimb/src/main/java/com/youngclimb/domain/model/repository/CenterRepository.java
@@ -3,5 +3,5 @@ package com.youngclimb.domain.model.repository;
 import com.youngclimb.domain.model.entity.Center;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CenterRepository extends JpaRepository<Center, Long> {
+public interface CenterRepository extends JpaRepository<Center, Integer> {
 }
