+++ BackEnd/youngclimb/src/main/java/com/youngclimb/domain/model/repository/WallRepository.java
@@ -3,5 +3,5 @@ package com.youngclimb.domain.model.repository;
 import com.youngclimb.domain.model.entity.Wall;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WallRepository extends JpaRepository<Wall, Long> {
+public interface WallRepository extends JpaRepository<Wall, Integer> {
 }
